kvs v0.1.0 kvs

Types

keyword/0

keyword() :: [{atom(), term()}].

An Elixir-compatible keyword list. A list of pairs in which the first item (the key) is an atom.

t/0

t() :: [{term(), term()}].

The default kvs type - a list of key-value pairs.

Functions

from_proplist/1
(since 0.1.0)

from_proplist(proplists:proplist()) -> t().

Unfolds a Proplist into a list of key-value pairs.

Properties which are not pairs are substituted with pairs where the second element of each is true. The call fails with a function_clause if Proplist is not a list.

get/2
(since 0.1.0)

get(Key, KVs) -> Value
       when Key :: term(), KVs :: t(), Value :: term().

Returns the Value associated with Key if KVs contains Key.

The call fails with a {badkey, Key} exception if no value is associated with Key, or with a function_clause if KVs is not a list.

Example:

  > kvs:get(a, [{a, 3}]).
  3
  > kvs:get(z, [{a, 3}]).
  ** exception error: {badkey,z}
       in function  kvs:get/2
          called as kvs:get(z,[{a,3}])

get/3
(since 0.1.0)

get(Key, KVs, Default) -> Value | Default
       when
           Key :: term(),
           KVs :: t(),
           Default :: term(),
           Value :: term().

Returns Value associated with Key if KVs contains Key. If no value is associated with Key, Default is returned.

The call fails with a function_clause if KVs is not a list.

Example:

  > kvs:get(a, [{a, "Just a"}], "Default value").
  "Just a"
  > kvs:get(z, [{a, "Just a"}], "Default value").
  "Default value"

proplist_to_map/1
(since 0.1.0)

proplist_to_map(proplists:proplist()) -> map().

Returns a map built from Proplist.

The proplist is first unfolded, i.e. properties which are not pairs are substituted with pairs where the second element of each is true. The call fails with a function_clause if Proplist is not a list.