kvs (kvs v0.1.0)

A fail-fast interface to key-value lists.

Types

keyword()

Specs

keyword() :: [{atom(), any()}]

An Elixir-compatible keyword list. A list of pairs in which the first item (the key) is an atom.

keyword(v)

Specs

keyword(v) :: [{atom(), v}]

An Elixir-compatible keyword list. A list of pairs in which the first item (the key) is an atom and the second item is of type V.

t()

Specs

t() :: [{any(), any()}]

The default kvs type - a list of key-value pairs.

t(k, v)

Specs

t(k, v) :: [{k, v}]

A list of key-value pairs, where the keys are of type K and values are of type V.

Functions

from_proplist/1
(since 0.1.0)

Specs

from_proplist(:proplists.proplist()) :: t()

Unfolds a Proplist into a list of key-value pairs.

Properties which are not pairs are substituted with pairs where the second element of each is true.

The call fails with a function_clause if Proplist is not a list.

Example:

  erl> kvs:from_proplist([a, {b, 2}]).
  [{a,true},{b,2}]
  erl> kvs:from_proplist(a).
  ** exception error: no function clause matching
                      kvs:from_proplist(a)
                      (/Users/erszcz/work/erszcz/kvs/src/kvs.erl, line 68)

get/2
(since 0.1.0)

Specs

get(key, t(key, value)) :: value when key: var, value: var

Returns the Value associated with Key if KVs contains Key.

The call fails with a {badkey, Key} exception if no value is associated with Key, or with a function_clause if KVs is not a list.

Example:

  erl> kvs:get(a, [{a, 3}]).
  3
  erl> kvs:get(z, [{a, 3}]).
  ** exception error: {badkey,z}
       in function  kvs:get/2
          called as kvs:get(z,[{a,3}])

get/3
(since 0.1.0)

Specs

get(key, t(key, value), default) :: value | default
when key: var, default: var, value: var

Returns Value associated with Key if KVs contains Key. If no value is associated with Key, Default is returned.

The call fails with a function_clause if KVs is not a list.

Example:

  erl> kvs:get(a, [{a, "Just a"}], "Default value").
  "Just a"
  erl> kvs:get(z, [{a, "Just a"}], "Default value").
  "Default value"

proplist_to_map/1
(since 0.1.0)

Specs

proplist_to_map(:proplists.proplist()) :: map()

Returns a map built from Proplist.

The proplist is first unfolded, i.e. properties which are not pairs are substituted with pairs where the second element of each is true.

The call fails with a function_clause if Proplist is not a list.

Example:

  erl> kvs:proplist_to_map([a, {b, 2}]).
  #{a => true,b => 2}
  erl> kvs:proplist_to_map(a).
  ** exception error: no function clause matching
                      kvs:proplist_to_map(a)
                      (/Users/erszcz/work/erszcz/kvs/src/kvs.erl, line 125)